+++ testing/lutmessage.js
@@ -836,6 +836,7 @@ LUTMessage.prototype.loadGamutLUTs = function() {
 	var fileNames = [
 		'LC709',
 		'LC709A',
+		's709',
 		'cpouttungsten',
 		'cpoutdaylight',
 		'Amira709',